Services y Test

+++ src/main/java/edu/icesi/fitness/model/User.java
@@ -9,7 +9,7 @@ import java.util.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "email", nullable = false)
     private String email;
@@ -71,7 +71,7 @@ public class User {
 
     public User() {}
 
-    public Integer getId(){return id;} public void setId(Integer id){this.id=id;}
+    public Long getId(){return id;} public void setId(Long id){this.id=id;}
     public String getName(){return name;} public void setName(String name){this.name=name;}
     public Integer getAge(){return age;} public void setAge(Integer age){this.age=age;}
     public Double getWeight(){return weight;} public void setWeight(Double weight){this.weight=weight;}
